Segregating models to follow Single Responsibility Principle
