+++ ExamApp/Context/MainContext.cs
@@ -37,26 +37,3 @@ public class MainContext : DbContext
     }
 }
 
-public class Course
-{
-    [Key]
-    public Guid Id { get; set; }
-    public string Title { get; set; }
-    public string Description { get; set; }
-
-    public virtual ICollection<Student> Students { get; set; }
-    public virtual Language Language { get; set; }
-}
-
-public class Student
-{
-    [Key]
-    public int Id { get; set; }
-    public string Name { get; set; }
-    public int Age { get; set; }
-    public Guid CourseId { get; set; }
-
-    public virtual Course Course { get; set; }
-}
-
-public record Language(Guid Id, string Title);
