+++ ExamApp/ExamApp.csproj
@@ -10,5 +10,9 @@
       <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="6.0.3" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Model\" />
+    </ItemGroup>
+
 
 </Project>
