+++ ExamApp/Model/Language.cs
@@ -1,0 +1,6 @@
+﻿using System;
+
+namespace ExamApp
+{
+    public record Language(Guid Id, string Title);
+}
